# vodacom-github/target/generated-sources/openapi/src/main/java/za/vodacom/repoprofile/api/ProfilesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package za.vodacom.repoprofile.api;

import za.vodacom.repoprofile.dto.ErrorResponse;
import za.vodacom.repoprofile.dto.UserProfileResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-16T17:49:49.951314600+02:00[Africa/Johannesburg]")
@Validated
@Tag(name = "profiles", description = "the profiles API")
public interface ProfilesApi {

    default ProfilesApiDelegate getDelegate() {
        return new ProfilesApiDelegate() {};
    }

    /**
     * POST /profiles/{username} : Retrieve user profile and repositories
     * Retrieve a user’s profile details along with their repositories sorted by stargazers count, including language breakdowns.
     *
     * @param username  (required)
     * @return Successful response (status code 200)
     *         or Bad Request – invalid or missing username (status code 400)
     *         or Internal Server Error – unexpected failure when querying GitHub (status code 500)
     */
    @Operation(
        operationId = "getProfile",
        summary = "Retrieve user profile and repositories",
        description = "Retrieve a user’s profile details along with their repositories sorted by stargazers count, including language breakdowns.",
        tags = { "profiles" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserProfileResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request – invalid or missing username", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error – unexpected failure when querying GitHub", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/profiles/{username}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserProfileResponse> getProfile(
        @Parameter(name = "username", description = "", required = true, in = ParameterIn.PATH) @PathVariable("username") String username
    ) {
        return getDelegate().getProfile(username);
    }

}